refactored library
