+++ myPortfolioManagement/MeteoraGlobalEquityFund/Archive/fx_impact_on_portfolio.py
