+++ myPortfolioManagement/MeteoraGlobalEquityFund/Archive/get_weathlist_data.py
